Fix parent child routing strategy
