+++ src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { FormsDashboardComponent } from './forms/forms-dashboard/forms-dashboard
 import { LoginComponent } from './login/login.component';
 import { RolesComponent } from './roles/roles.component';
 import { CreateUserComponent } from './users/create-user/create-user.component';
+import { UsersInfoComponent } from './users/users-info/users-info.component';
 import { UsersComponent } from './users/users.component';
 import { WelcomeComponent } from './welcome/welcome.component';
 
@@ -26,15 +27,12 @@ const routes: Routes = [
   },
   {
     path: 'usersParent', children: [
-      { path: 'users', component: UsersComponent },
+      { path: 'users', component: UsersInfoComponent, },
       { path: 'roles', component: RolesComponent },
-    ], component: UsersComponent,
+
+    ],
   },
 
-  {
-    path: 'usersParent/createUser',
-    component: CreateUserComponent
-  },
   { path: '', component: LoginComponent },
   { path: '**', component: AppComponent }
 ];
